Fixed LoginInstrumentedTest and added UserProfileActivity
The loginInstrumentedTest all pass but still need to add some more. UserProfileActivity added: on login, we check if the SharedPreferences file already exists, if not, user is directed to a new page to enter his infos which are saved to a new SharedPreferences file. If file already exist, login proceeds directly to MapsActivity.