Strugling with adding fcm admin sdk
Since it would be too hard to find the difference between to successive snapshot of the database to notify the right user that a participant is coming, I am currently trying another idea: I added a token fiel to Need corresponding to the user/app instance that created the need. Now when a participant is added to a need, I just have to send a message to the device identified by that token. Problem with initializing admin sdk, how do you do that shit